pyproject.toml
[tool.poetry]
name = "magenta-rapids-api"
version = "0.1.0"
description = ""
authors = ["Aaron Grisez <acgrisez@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.1"
fastapi = {extras = ["all"], version = "^0.54.1"}
pydantic = {extras = ["dotenv"], version = "^1.6.1"}
sqlalchemy = "^1.3.16"
soundfile = "^0.10.3"
schedule = "^0.6.0"
jaeger-client = "^4.3.0"
semver = "^2.10.1"
aiocache = {extras = ["redis"], version = "^0.11.1"}
watchdog = {extras = ["watchmedo"], version = "^0.10.3"}
contextvars = "^2.4"
pymongo = "^3.11.2"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pytest = "^5.4.1"
flake8 = "^3.7.9"
mypy = "^0.770"
autoflake = "^1.3.1"
coverage = "^5.1"
pre-commit = "^2.3.0"
datamodel-code-generator = "^0.6.11"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
